backend/app/requirements.txt
gunicorn == 20.1.0
django == 4.0.1
django-dotenv == 1.4.2
djangorestframework == 3.13.1
django-filter == 21.1
djangorestframework-simplejwt == 5.0.0
django-allauth == 0.47.0
dj-rest-auth == 2.2.2
psycopg2 == 2.9.3
django-seed == 0.3.1 
Pillow == 9.0.0
drf-spectacular == 0.21.1
requests == 2.27.1